converting Carousel to tsx
