+++ src/components/Carousel/Carousel-Item.model.ts
@@ -1,0 +1,5 @@
+export interface Item {
+    name: string;
+    description: string;
+    imgSrc?: string;
+}
