+++ src/pages/LandingPage.tsx
@@ -2,8 +2,10 @@ import React from "react";
 import { Box } from "@material-ui/core";
 import Header from "../components/Header";
 import Map from "../components/MapBanner";
-import CityCarousel from "../components/Carousel";
+import CityCarousel from "../components/Carousel/Carousel";
 import CityTable from "../components/Table";
+
+
 const LandingPage: React.FC = () => {
   return (
     <Box>
